+++ packages/moonstone/CHANGELOG.md
@@ -45,6 +45,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/IncrementSlider` so that the knob is spottable via pointer, and 5-way navigation between the knob and the increment/decrement buttons is functional
 - `moonstone/Slider` and `moonstone/IncrementSlider` to not fire `onChange` for value changes from props
 - `spotlightDisabled` property support for spottable moonstone components
+- `moonstone/Popup` and `moonstone/ContextualPopupDecorator` so that when the popup is closed, spotlight focus returns to the control that had focus prior to the popup opening.
 
 ## [1.0.0-beta.1] - 2016-12-30